Rename pokedex Django app to backend
